+++ src/components/common/sidebar/styles.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import theme from '../../theme';
+import theme from '../../../theme';
 
 export const Container = styled.div`
