+++ src/components/main/styles.tsx
