--- src/components/questionTile/index.tsx
@@ -1,20 +1,0 @@
-import React from 'react';
-import { Container, Wrapper } from './styles';
-
-interface Props {
-    filter: string,   
-}
-const QuestionTileContainer: React.FC<Props> = ({filter}:Props) => {
-
-    return (
-
-        <Container>
-            <Wrapper>
-                
-            </Wrapper>
-        </Container>
-
-    );
-};
-
-export default QuestionTileContainer;
