--- src/components/questionTile/item/tile/QuestionTile.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import { Container } from './styles';
-
-const QuestionTile: React.FC = () => {
-    return(
-        <Container>
-            
-        </Container>
-    );
-};
-
-export default QuestionTile;
