--- src/components/questionTile/item/tile/styles.tsx
@@ -1,4 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.div`
-`;
