--- src/components/questionTile/styles.tsx
@@ -1,6 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.div`
-`;
-
-export const Wrapper = styled.div``;
