+++ src/routes/main/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import MainComponent from "../../components/mainPage";
+import MainComponent from "../../components/main";
 
 const Main: React.FC = () => <MainComponent />;
